refactor: change sendCurrentRate method due to rethinking of solid priciples
