+++ src/email/interfaces/index.ts
@@ -1,0 +1,1 @@
+export * from './send-email-options.interface';
