+++ src/rate/interfaces/exchange-rate.interface.ts
@@ -1,7 +1,4 @@
 export interface IExchangeRate {
-  r030: number;
-  txt: string;
   rate: number;
-  cc: string;
   exchangedate: string;
 }
