+++ src/rate/interfaces/get-exchange-rate.interface.ts
@@ -1,0 +1,7 @@
+export interface IGetExchangeRate {
+  r030: number;
+  txt: string;
+  rate: number;
+  cc: string;
+  exchangedate: string;
+}
