+++ src/rate/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from './get-exchange-rate.interface';
+export * from './exchange-rate.interface';
