+++ src/rate/rate.controller.ts
@@ -1,14 +1,14 @@
 import { Controller, Get } from '@nestjs/common';
 
+import { IExchangeRate } from './interfaces';
 import { RateService } from './rate.service';
-import { CurrentRateResponse } from './responses/current-rate.response';
 
 @Controller('rate')
 export class RateController {
   constructor(private readonly rateService: RateService) {}
 
   @Get()
-  async getCurrentRate(): Promise<CurrentRateResponse> {
+  async getCurrentRate(): Promise<IExchangeRate> {
     return await this.rateService.getCurrentRate();
   }
 }
