--- src/rate/responses/current-rate.response.ts
@@ -1,4 +1,0 @@
-export class CurrentRateResponse {
-  rate: number;
-  exchangedate: string;
-}
