+++ src/task-schedule/task-schedule.service.ts
@@ -3,13 +3,19 @@ import { Cron, CronExpression } from '@nestjs/schedule';
 
 import { TIMEZONE } from './task-schedule.constants';
 import { EmailService } from '../email/services/email.service';
+import { RateService } from '../rate/rate.service';
 
 @Injectable()
 export class TaskScheduleService {
-  constructor(private readonly emailService: EmailService) {}
+  constructor(
+    private readonly emailService: EmailService,
+    private readonly rateService: RateService,
+  ) {}
 
   @Cron(CronExpression.EVERY_DAY_AT_10AM, { timeZone: TIMEZONE })
   async sendCurrentRateEmail(): Promise<void> {
-    await this.emailService.sendCurrentRate();
+    const rate = await this.rateService.getCurrentRate();
+
+    await this.emailService.sendCurrentRate(rate);
   }
 }
